+++ package.json
@@ -113,17 +113,17 @@
     "gerror": "^1.0.2",
     "json-rpc-peer": "^0.17.0",
     "open-graph": "^0.2.6",
-    "state-switch": "^1.1.15",
+    "state-switch": "^1.2.1",
     "uuid": "^8.3.2",
-    "wechaty-puppet": "^1.2.7",
+    "wechaty-puppet": "^1.2.11",
     "wechaty-puppet-service": "^1.0.9",
     "ws": "^8.2.3"
   },
   "devDependencies": {
-    "@chatie/eslint-config": "^1.0.3",
+    "@chatie/eslint-config": "^1.0.4",
     "@chatie/git-scripts": "^0.6.2",
     "@chatie/semver": "^0.4.7",
-    "@chatie/tsconfig": "^1.0.2",
+    "@chatie/tsconfig": "^1.0.3",
     "@types/dotenv": "^8.2.0",
     "@types/glob": "^7.2.0",
     "@types/open-graph": "^0.2.2",
@@ -140,7 +140,7 @@
     "sloc": "^0.2.1",
     "typed-emitter": "^1.4.0",
     "typedoc": "^1.0.0-dev.4 ",
-    "wechaty-puppet-mock": "^1.0.1",
+    "wechaty-puppet-mock": "^1.0.3",
     "wechaty-puppet-padlocal": "^0.4.2"
   },
   "files": [
