+++ src/mods/users.ts
@@ -33,4 +33,4 @@ export type {
   Delay,
   Tag,
   UrlLink,
-}                   from '../user/mod.js'
+}                   from '../user-modules/mod.js'
