+++ src/user-modules/friendship.ts
@@ -39,8 +39,8 @@ import type {
 }                       from './room.js'
 import {
   wechatifyMixin,
-}                       from './mixins/wechatify.js'
-import { validationMixin } from './mixins/validation.js'
+}                       from '../user-mixins/wechatify.js'
+import { validationMixin } from '../user-mixins/validation.js'
 
 interface FriendshipAddOptionsObject {
   room?: Room,
