+++ src/user-modules/mini-program.ts
@@ -22,12 +22,12 @@ import type * as PUPPET  from 'wechaty-puppet'
 import type { Constructor } from '../deprecated/clone-class.js'
 import { log } from '../config.js'
 
-import { validationMixin } from './mixins/validation.js'
+import { validationMixin } from '../user-mixins/validation.js'
 
 import {
   EmptyBase,
   wechatifyMixin,
-}                       from './mixins/wechatify.js'
+}                       from '../user-mixins/wechatify.js'
 
 const MixinBase = wechatifyMixin(
   EmptyBase,
